# q_shared: inline some vector functions