… src/engine/qcommon/q_math.cpp