This helps to get more performance in some place of the code.

One good example that benefits from it is the `R_TBNtoQtangents()` function. This function is used for rendering models when the models is too large for some GPU limit. It's possible to force the engine to use this code path for all models with `r_vboVertexSkinning 0`. A good example of scene that puts the engine on knees is the `plat23` map with `setviewpos 1920 1920 0 0 0`.

This simple change increases performance by `15%` on this heavy scene, with the `ondemand` CPU profile, it makes the game run at `77fps` instead of  `67fps` on my machine.

Before:

![unvanquished_2024-09-14_035249_000](https://github.com/user-attachments/assets/128012c8-2a9d-404a-b7ba-6b161ab5f5ac)

After:

![unvanquished_2024-09-14_035455_000](https://github.com/user-attachments/assets/f36aec35-bb67-4e77-9677-8732a983208f)

When I set the CPU profile to `performance` and do some tricks like suspending other apps and setting `-20` priority to the renderer thread, that patch makes the game run at `83fps` instead of `72fps` on my machine.

Before:

![unvanquished_2024-09-14_035903_000](https://github.com/user-attachments/assets/4d102e6a-1978-4682-81fe-8f015baaa611)

After:

![unvanquished_2024-09-14_040318_000](https://github.com/user-attachments/assets/058759f5-b944-4558-af8a-5787f8435068)