@illwieckz 2024-09-14
I can reproduce the framerate difference very consistently, not only the framerate is stable and lasts for minutes or hours during tests, but I get consistently the same result if when re-run the test multiple times in a row, and when I re-run the test hours later.