@illwieckz 2024-09-14
@VReaperV is the CI build error in Material uncovered by this change?

- https://dev.azure.com/UnvanquishedDevelopment/Daemon/_build/results?buildId=9738&view=logs&jobId=e4954a55-bc0b-5fbe-d8f6-05775b1a9559&j=e4954a55-bc0b-5fbe-d8f6-05775b1a9559&t=6fedcdd4-7b5b-58c7-109c-77b7f76a6cd5