@VReaperV 2024-09-14
> @VReaperV is the CI build error in Material uncovered by this change?

Yep, needs a `vec3_t` somewhere in `MaterialSystem` for PVS lock.