@illwieckz 2024-09-14
For now I `VectorClear()` it, it seems to make the CI happy.